Split off functionality into separate files in includes/; added placeholder 404 template
