+++ style.css
@@ -1,74 +1,6 @@
 /*
 Theme Name: Good Morning UCF WordPress Theme
 Author: UCF Web Communications
-Github Theme URI: https://github.com/UCF/GMUCF
-Version: v2.1.7
+Github Theme URI: https://github.com/UCF/UCF-GMUCF-Theme
+Version: v3.0.0
 */
-
-/*/ Global Styles /*/
-html {background:none;}
-#blueprint-container {width:600px;margin:auto;}
-#UCFHBHeader {display:none;}
-.narrow {}
-
-h1 {font-size:45px;font-weight:300;margin-bottom:20px;}
-
-/* Header */
-#header {border-bottom:1px solid #ddd;padding-bottom:20px;height:50px;}
-#header > div {float:left;}
-#header .weather {}
-#header .weather,
-	#header #date {border-right:1px solid #CCC;padding:0 20px;height:60px;}
-#header .weather {font-size:16px;}
-#header .weather img {position:relative;top:10px;}
-#header .weather .temp {font-weight:bold;}
-/*
-#header .weather img {float:left;margin-right:15px;}
-#header .weather .temp {font-size:25px;margin-top:10px;text-align:center;}
-#header .weather .type {font-size:20px;font-weight:100;}
-*/
-#header #date {padding-left:0;text-align:center;font-weight:200;font-size:20px;font-family:"Helvetica Neue",Helvetica,sans-serif;}
-#header #date #day {position:relative;top:20px;}
-#header #date #month {position:relative;top:20px;}
-#header #this-day-in-history {width:90px;float:left;font-weight:normal;display:block;padding-top:20px;padding-left:15px;font-family:"Helvetica Neue",Helvetica,sans-serif;font-size:14px;text-decoration:none;}
-
-/* Top Story */
-#top_story h2 {margin-bottom:0;}
-#top_story h2 a {font-family:Helvetica,sans-serif;font-size:32px;margin-bottom:0;font-weight:100;}
-#top_story p a {color:#9D1A1A;text-decoration:underline;font-family:Helvetica,sans-serif;font-size:17px;font-weight:100;font-size:18px;line-height:1em;}
-#top_story p a:first-child {text-decoration:none;color:#333;}
-
-
-#bottom #ucf {width:330px;padding-left:15px;float:left;border-top:1px solid #ddd;padding-top:15px;}
-#bottom #social {width:185px;float:right;border-top:1px solid #ddd;padding-top:15px;margin-top:7px;}
-
-/* Featured Stories */
-.featured-story {width:300px;float:left;}
-.featured-story img {float:left;}
-.featured-story h3 {font-family: Helvetica, sans-serif;font-size:14px;margin-top:0;}
-.featured-story p {font-family: Helvetica, sans-serif;font-size:14px;margin-top:0;line-height:1.3em}
-.featured-story .content {font-size:14px;padding-left:110px;padding-right:10px;font-weight:100;}
-
-/* Announcemnts */
-#announcements {}
-#announcements h2 {font-weight:100;font-family:Helvetica,sans-serif;}
-#announcements > ul {margin:0;padding:0;}
-#announcements > ul > li {list-style-type:none;}
-#announcements > ul > li > a {color:#9D1A1A;font-weight:100;font-family:Helvetica,sans-serif;}
-#announcements > p { text-align: right; }
-#announcements > p  > a { text-decoration: underline; color:#9D1A1A;font-weight:100;font-family:Helvetica,sans-serif; }
-
-/* Bottom */
-#bottom {clear:both;margin-top:40px;border-top:1px solid #ddd;padding-top:15px;margin-bottom:30px;}
-#bottom #social,
-	#bottom #ucf {border:none;padding:0;}
-#bottom #social h2 {font-family:Helvetica,sans-serif;font-weight:100;margin-bottom:5px;}
-#bottom #social a {text-indent:-300em;width:32px;height:32px;display:block;float:left;margin-left:10px;}
-#bottom #social #facebook {background:#FFF url('static/img/social/facebook.png') center center no-repeat;margin-left:0;}
-#bottom #social #youtube {background:#FFF url('static/img/social/youtube.png') center center no-repeat;}
-#bottom #social #twitter {background:#FFF url('static/img/social/twitter.png') center center no-repeat;}
-#bottom #social #instagram {background:#FFF url('static/img/social/instagram.png') center center no-repeat;}
-#bottom #ucf {display:block;background:#FFF url('static/img/logo-no-opportunity.png') top left no-repeat;font-family:Helvetica,sans-serif;text-decoration:none;padding-top:61px;min-width:152px;}
-
-/* Footer */
-#footer {display:none;}
